// Frontend/src/app/modules/assessments/models/writing.assessmen.item.model.ts
import { ImageFile } from '../../../@shared/components/file-loader/file-loader.component';
import { WritingAssessmentItemState } from './writing.assessment.item.state.model';
import { Writing } from './writing.model';

export interface WritingAssessmentItem {
  student_id: string;
  student_name: string;
  file: ImageFile;
  state: WritingAssessmentItemState;
  writing: Writing;
}